F/succinct trie (#10)
* Move things around

* First version of succinct trie

* Fix prefix filtering in succinct trie

+++ bench/SuRFFacade.h
@@ -10,7 +10,7 @@ class SuRFFacade : public PrefixFilter {
 public:
     explicit SuRFFacade(std::vector<std::string> &keys, bool real, uint32_t bits_cnt);
     bool lookupPrefix(const std::string &prefix) override;
-    unsigned long long getMemoryUsage() const override;
+    uint64_t getMemoryUsage() const override;
     std::string getName() const override { return name_; }
 private:
     surf::SuRF *surf_;
@@ -34,7 +34,7 @@ bool SuRFFacade::lookupPrefix(const std::string &prefix) {
     return surf_->lookupRange(prefix, true, right_key, false);
 }
 
-unsigned long long SuRFFacade::getMemoryUsage() const {
+uint64_t SuRFFacade::getMemoryUsage() const {
     return surf_->getMemoryUsage();
 }
 }
